+++ t/030-OT-File.t
@@ -12,19 +12,19 @@ subtest 'OT File', {
 
   $f .= new( :object<t/030-OT-File.t>, :type(OT-File));
   my BSON::Document $d = $f.meta;
-  is $d<object-name>, '030-OT-File.t', $d<object-name>;
-  is $d<object-extension>, 't', $d<object-extension>;
-  like $d<object-path>, /:s t $/, $d<object-path>;
-  ok $d<object-exists>, 'object exists';
+  is $d<name>, '030-OT-File.t', $d<name>;
+  is $d<extension>, 't', $d<extension>;
+  like $d<path>, /:s t $/, $d<path>;
+  ok $d<exists>, 'object exists';
   ok $d<content-sha1>, 'sha calculated on content';
 
 #  say $d.perl;
 
   $f .= new( :object<t/other-file.t>, :type(OT-File));
   $d = $f.meta;
-  is $d<object-name>, 'other-file.t', $d<object-name>;
-  like $d<object-path>, /:s t $/, $d<object-path>;
-  ok !$d<object-exists>, 'object does not exist';
+  is $d<name>, 'other-file.t', $d<name>;
+  like $d<path>, /:s t $/, $d<path>;
+  ok !$d<exists>, 'object does not exist';
   ok !$d<content-sha1>, 'no sha on content';
 
 #  say $d.perl;